% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms.R
\name{dtms}
\alias{dtms}
\title{Create dtms object}
\usage{
dtms(transient, absorbing, time, timestep = NULL)
}
\arguments{
\item{transient}{A character vector of names of the transient states in the state space}

\item{absorbing}{A character vector of names of the absorbing states in the state space}

\item{time}{A numeric vector with the time scale, including the starting time and the final time}

\item{timestep}{Step length of the time scale, will be guessed if NULL}
}
\value{
Returns an object of class 'dtms' to be passed to other functions of the package
}
\description{
Create dtms object
}
\examples{
dtms(transient=c("A","B"),
     absorbing="X",
     time=1:10)
}
