% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_boot.R
\name{dtms_boot}
\alias{dtms_boot}
\title{Bootstrap and block bootstrap}
\usage{
dtms_boot(
  data,
  dtms,
  fun,
  rep,
  method = "simple",
  idvar = "id",
  seed = NULL,
  verbose = FALSE,
  progress = FALSE,
  parallel = FALSE,
  cores = 2,
  .packages = c("mclogit", "VGAM", "nnet", "dtms"),
  ...
)
}
\arguments{
\item{data}{Data frame in transition format as created with \code{dtms_format}.}

\item{dtms}{dtms object, as created with \code{dtms}.}

\item{fun}{Function to be repeatedly applied, see details.}

\item{rep}{Numeric, number of bootstrap replications.}

\item{method}{Character (optional), either "simple" for simple bootstrap or "block" for block bootstrap. Default is "simple".}

\item{idvar}{Character (optional), name of ID variable in `data' identifying units. Only required for block bootstrap. Default is "id".}

\item{seed}{Numeric (optional), seed for random numbers. See details.}

\item{verbose}{Logical (optional), print output which might be generated when running `fun`? Default is FALSE.}

\item{progress}{Logical (optional), indicate progress if simple bootstrap? Default is FALSE.}

\item{parallel}{Logical (optional), use parallel processing? Default is FALSE.}

\item{cores}{Numeric (optional), if parallel=TRUE, how many cores should be used? Default is 2.}

\item{.packages}{Character (optional), packages to be loaded when parallel processing. Default is `c("mclogit","VGAM","nnet","dtms")`}

\item{...}{Arguments to be passed to `fun`, only works if `parallel=FALSE`.}
}
\value{
A list of results, see details
}
\description{
This function is a simple wrapper for bootstrapping and block-bootstrapping
data in transition format.
}
\details{
This is currently a placeholder. What goes here: fun, seed (if null), progress, parallel
}
\examples{
#Currently none
}
