% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_boot_summary.R
\name{dtms_boot_summary}
\alias{dtms_boot_summary}
\title{Summary function for bootstrap results}
\usage{
dtms_boot_summary(boot, probs = NULL, alpha = 0.05)
}
\arguments{
\item{boot}{Object created with \code{dtms_boot()}.}

\item{probs}{Numeric (optional), vector of percentiles. Default is NULL.}

\item{alpha}{Numeric (optional), confidence level. Default is 0.05.}
}
\value{
A list
}
\description{
Provides bootstrap percentiles for \code{dtms_boot()}.
}
\details{
Percentiles can be specified with the argument \code{probs}. If it is not
specified, the argument \code{alpha} is used instead. \code{alpha} is the
confidence level for
}
