% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_censoring.R
\name{dtms_censoring}
\alias{dtms_censoring}
\title{Get information on left censoring, right censoring, and gaps in data}
\usage{
dtms_censoring(
  data,
  dtms,
  fromvar = "from",
  tovar = "to",
  timevar = "time",
  idvar = "id",
  add = F,
  addtype = "id",
  print = T,
  printlong = F,
  varnames = c("LEFT", "GAP", "RIGHT")
)
}
\arguments{
\item{data}{Data frame}

\item{dtms}{dtms object}

\item{fromvar}{Variable with starting state}

\item{tovar}{Variable with receiving state}

\item{timevar}{Variable with time scale}

\item{idvar}{Variable with unit ID}

\item{add}{Add indicators to data set?}

\item{addtype}{If add=T, what type of information should be added? Either "id" or "obs".}

\item{print}{Print short table?}

\item{printlong}{Print long table?}

\item{varnames}{If variables are added, what should be their names}
}
\value{
Table or data frame
}
\description{
Get information on left censoring, right censoring, and gaps in data
}
