% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_clean.R
\name{dtms_clean}
\alias{dtms_clean}
\title{Cleans data in transition format}
\usage{
dtms_clean(
  data,
  dtms,
  fromvar = "from",
  tovar = "to",
  timevar = "time",
  dropTime = T,
  dropNA = T,
  dropAbs = T
)
}
\arguments{
\item{data}{Data frame (or similar) in transition format as created with `dtms_format`.}

\item{dtms}{DTMS object as created with `dtms`}

\item{fromvar}{Character string (optional), name of the variable with the starting state.}

\item{tovar}{Character string (optional), name of the variable with the receiving state.}

\item{timevar}{Character string (optional), name of the variable with the time scale.}

\item{dropTime}{Logical, drop observations with values of time not covered by the model. Default is TRUE.}

\item{dropNA}{Logical, drop observations with gaps, last observations, and similar. Default is TRUE.}

\item{dropAbs}{Logical, drop observations starting from absorbing states. Default is TRUE.}
}
\value{
Returns a cleaned data frame in transition format.
}
\description{
Cleans data in transition format. It can handle gaps in sequences and final observations,
observatiosn not covered by the time range, and observations starting in absorbing states.
}
\examples{
# Define model
simple <- dtms(transient=c("A","B"),
absorbing="X",
time=0:20)
# Transiton format
estdata <- dtms_format(data=simpledata,
dtms=simple,
idvar="id",
timevar="time",
statevar="state")
# Clean data
estdata <- dtms_clean(data=estdata,
                      dtms=simple)
}
