% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_data_summary.R
\name{dtms_data_summary}
\alias{dtms_data_summary}
\title{Summarize data in transition format}
\usage{
dtms_data_summary(
  data,
  dtms = NULL,
  fromvar = "from",
  tovar = "to",
  weightvar = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{dtms}{Dtms object}

\item{fromvar}{Variable with starting state}

\item{tovar}{Variable with receiving state}

\item{weightvar}{Variable with weigths}
}
\value{
A data frame with counts and probabilities
}
\description{
Summarize data in transition format
}
\examples{
simple <- dtms(transient=c("A","B"),
               absorbing="X",
               timescale=0:20)
estdata <- dtms_format(data=simpledata,
                       dtms=simple,
                       idvar="id",
                       timevar="time",
                       statevar="state")
dtms_data_summary(estdata)
}
