% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_duration.R
\name{dtms_duration}
\alias{dtms_duration}
\title{Generate variable with duration}
\usage{
dtms_duration(
  data,
  dtms,
  newname = "duration",
  fromvar = "from",
  idvar = "id",
  timevar = "time",
  ignoreleft = TRUE
)
}
\arguments{
\item{data}{A data set in transition format}

\item{dtms}{dtms object, as created with \code{dtms}.}

\item{newname}{Name of new variable}

\item{fromvar}{Character (optional), name of variable in `data` with starting state. Default is "from".}

\item{idvar}{Character (optional), name of variable in `data` with unit ID. Default is "id".}

\item{timevar}{Character (optional), name of variable in `data` with time scale. Default is "time".}

\item{ignoreleft}{Logical (optional), ignore left censoring and start counting at the first observation? Default is TRUE.}
}
\description{
Generate variable with duration
}
