% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_expectancy.R
\name{dtms_expectancy}
\alias{dtms_expectancy}
\title{Title}
\usage{
dtms_expectancy(
  matrix,
  transient = NULL,
  time = NULL,
  timestep = NULL,
  dtms = NULL,
  start_time = NULL,
  start_state = NULL,
  start_distr = NULL,
  end_time = NULL,
  sep = "_",
  correction = 0.5,
  total = T,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
