% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_fit.R
\name{dtms_fit}
\alias{dtms_fit}
\title{Title}
\usage{
dtms_fit(
  fromvar = "from",
  tovar = "to",
  timevar = "time",
  data,
  controls = NULL,
  formula = NULL,
  weights = NULL,
  reference = 1,
  method = "VGAM",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
