% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_plot.R
\name{dtms_plot}
\alias{dtms_plot}
\title{Plotting transition probabilities}
\usage{
dtms_plot(
  probs,
  dtms = NULL,
  fromvar = "from",
  tovar = "to",
  timevar = "time",
  Pvar = "P",
  sep = "_",
  ...
)
}
\arguments{
\item{probs}{Object with transition probabilities}

\item{dtms}{Optional, dtms object describing the model. If not provided states will be guessed from 'probs'.}

\item{fromvar}{Optional, character string indicating the variable with the starting state.}

\item{tovar}{Optional, character string indicating the variable with the receiving state.}

\item{timevar}{Optional, character string indicating the variable with the time scale.}

\item{Pvar}{Optional, character string indicating the variable with the transition probabilities.}

\item{sep}{Optional, character string with separator for long state names. Default is "_".}

\item{...}{Further arguments passed to plot().}
}
\value{
Sends a plot to the current device
}
\description{
A simple function for plotting transition probabilities with base R. This is
fast, but it is much easier to produce nicer looking results with
dtms_simplify.
}
\examples{
## Model setup
simple <- dtms(transient=c("A","B"),
               absorbing="X",
               timescale=0:20)
## Reshape to transition format
estdata <- dtms_format(data=simpledata,
                       dtms=simple,
                       idvar="id",
                       timevar="time",
                       statevar="state")
## Clean
estdata <- dtms_clean(data=estdata,
                      dtms=simple)
## Fit model
fit <- dtms_fit(data=estdata)
## Predict probabilities
probs    <- dtms_transitions(dtms=simple,
                             model = fit)
## Plot
dtms_plot(probs=probs,
          dtms=simple)
}
