% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_risk.R
\name{dtms_risk}
\alias{dtms_risk}
\title{Title}
\usage{
dtms_risk(
  matrix,
  transient = NULL,
  time = NULL,
  dtms = NULL,
  risk,
  start_time = NULL,
  start_state = NULL,
  start_distr = NULL,
  end_time = NULL,
  sep = "_"
)
}
\arguments{
\item{sep}{}
}
\description{
Title
}
