% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_simplify.R
\name{dtms_simplify}
\alias{dtms_simplify}
\title{Simplify state names}
\usage{
dtms_simplify(probs, fromvar = "from", tovar = "to", sep = "_")
}
\arguments{
\item{probs}{Object with transition probabilities}

\item{fromvar}{Name of variable with starting state}

\item{tovar}{Name of variable with receiving state}

\item{sep}{Separator for long state name}
}
\value{
Data frame
}
\description{
Simplify state names
}
\examples{
## Define model: Absorbing and transient states, time scale
simple <- dtms(transient=c("A","B"),
               absorbing="X",
               timescale=0:20)
## Reshape to transition format
estdata <- dtms_format(data=simpledata,
                       dtms=simple,
                       idvar="id",
                       timevar="time",
                       statevar="state")
## Clean
estdata <- dtms_clean(data=estdata,
                      dtms=simple)
## Fit model
fit <- dtms_fit(data=estdata)
## Predict probabilities
probs    <- dtms_transitions(dtms=simple,
                             model = fit)
## Simplify
probs |>  dtms_simplify()
## NOT RUN: requires ggplot2
# library(ggplot2)
# probs |>  dtms_simplify() |>
#   ggplot(aes(x=time,y=P,color=to)) +
#   geom_line() +
#   facet_wrap(~from)
}
