% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_source.R
\name{dtms_source}
\alias{dtms_source}
\title{Source code only from specific lines}
\source{
User Matthew Plourde on Stackoverflow https://stackoverflow.com/questions/12214963/source-only-part-of-a-file
}
\usage{
dtms_source(file, start, end, local = FALSE)
}
\arguments{
\item{file}{Name of the file to read (character)}

\item{start}{First line to read (numeric)}

\item{end}{Last line to read (numeric)}

\item{local}{Logical or an environment. Default is 'FALSE'. See function \code{source} for details.}
}
\description{
This is a helper function which allows to source only specific lines from
a file.
}
\seealso{
[func(source)]
}
