% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_start.R
\name{dtms_start}
\alias{dtms_start}
\title{Title}
\usage{
dtms_start(
  transient = NULL,
  time = NULL,
  dtms = NULL,
  data,
  start_time = NULL,
  start_state = NULL,
  fromvar = "from",
  timevar = "time",
  variables = NULL,
  sep = "_"
)
}
\arguments{
\item{sep}{}
}
\description{
Title
}
