% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_transitions.R
\name{dtms_transitions}
\alias{dtms_transitions}
\title{Title}
\usage{
dtms_transitions(
  transient = NULL,
  absorbing = NULL,
  time = NULL,
  dtms = NULL,
  model,
  constant = NULL,
  varying = NULL,
  sep = "_",
  timestep = NULL,
  timevar = "time",
  fromvar = "from",
  tovar = "to",
  Pvar = "P"
)
}
\arguments{
\item{Pvar}{}
}
\description{
Title
}
