% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_visits.R
\name{dtms_visits}
\alias{dtms_visits}
\title{Title}
\usage{
dtms_visits(
  matrix,
  transient = NULL,
  time = NULL,
  timestep = NULL,
  dtms = NULL,
  risk,
  start_time = NULL,
  start_state = NULL,
  start_distr = NULL,
  end_time = NULL,
  method = "mid",
  sep = "_",
  total = F
)
}
\arguments{
\item{total}{}
}
\description{
Title
}
