% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_data.R
\docType{data}
\name{simple}
\alias{simple}
\title{simple: an artificial dataset with abstract trajectories}
\format{
`simple`
A data frame with 10,529 rows and 3 columns:
\describe{
  \item{id}{Identifier of the units}
  \item{age}{Age}
  \item{state}{The state occupied by an unit at a given age}
}
}
\usage{
simple
}
\description{
An artificial dataset with abstract states and time scale. The
state space consists of two transient states (A,B) and one absorbing state (X).
}
\keyword{datasets}
