% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_data.R
\docType{data}
\name{workdata}
\alias{workdata}
\title{workdata: simulated working trajectories}
\format{
`workdata`
A data frame with 250,000 rows and 4 columns:
\describe{
  \item{ID}{Person identifier}
  \item{Gender}{Individuals' gender (0=men, 1=women)}
  \item{Age}{Age, the time scale of this example}
  \item{State}{The state occupied by an unit at a given age, either `Working`, `Retired`, `Non-working`, or `Dead`.}
}
}
\source{
<https://doi.org/10.1007/s13524-017-0619-6>
}
\usage{
workdata
}
\description{
A simulated data set of individuals' working trajectories during late working life and retirement
age. The state space consists of three transient states (working; retired; not working)
and one absorbing state (dead). The age range covers ages 50 to 99. The data is
simulated using transition probabilities published as part of Dudel & Myrskylä (2017).
}
\keyword{datasets}
