% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_backward.R
\name{dtms_backward}
\alias{dtms_backward}
\title{Carry states backward}
\usage{
dtms_backward(
  data,
  state,
  statevar = "state",
  idvar = "id",
  timevar = "time",
  dtms = NULL,
  overwrite = "missing",
  vector = FALSE
)
}
\arguments{
\item{data}{A data frame in long format.}

\item{state}{Character, name of the state to be carried backward}

\item{statevar}{Character (optional), name of the variable in the data frame with the states. Default is "state".}

\item{idvar}{Character (optional), name of variable with unit ID. Default is `id`.}

\item{timevar}{Character (optional), name of variable with time scale. Default is `time`.}

\item{dtms}{dtms object (optional), as created with \code{dtms}. Not required if `overwrite==transient`.}

\item{overwrite}{Character (optional), one of `transient`, `missing`, `absorbing`, and `all`, see details. Default is `transient`.}

\item{vector}{Logical (optional), return vector (if TRUE) or data frame (if FALSE). Default is FALSE}
}
\value{
The data frame specified with `data` and the edited state variable (if `vector=FALSE`) or a vector (if `vector=TRUE`).
}
\description{
This function carries a state backward after its first occurrence.
}
\details{
This function carries a state backward after its first occurrence.
For instance, carrying the state "A" backward in the sequence `B, B, A, B, B`
will give the sequence `A, A, A, B, B`.

The data frame has to be in long format and not in transition format; i.e.,
usually, this function will be applied before \code{dtms_format}.

The argument `overwrite` is used to control what type of information is
replaced. If `overwrite==transient`, then only transient states are replaced
while missing values and absorbing states remain unchanged. For example,
carrying backward state "A" in the sequence `B, NA, B, B, X, A, X` with X
being an absorbing state will give `A, NA, A, A, X, A, X`. If
`overwrite==missing` then in addition to transient states also missing values
are replaced and for the example sequence `A, A, A, A, X, A, X` would be
returned. If `overwrite==absorbing` then in addition to transient states
absorbing states will be replaced; for the example sequence the result would
be `A, NA, A, A, A, A, X`. Finally, if `overwrite==all` then all values in
the sequence will be replaced: `A, A, A, A, A, A, X`.
}
\examples{
simple <- dtms(transient=c("A","B"),
absorbing="X",
timescale=0:19)

dtms_backward(data=simpledata,
             state="A",
             dtms=simple,
             overwrite="transient")
}
\seealso{
[func(dtms_forward)]
}
