% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_first.R
\name{dtms_first}
\alias{dtms_first}
\title{Title}
\usage{
dtms_first(
  matrix,
  transient = NULL,
  time = NULL,
  timestep = NULL,
  dtms = NULL,
  risk,
  start_time = NULL,
  start_state = NULL,
  start_distr = NULL,
  end_time = NULL,
  method = "mid",
  sep = "_",
  total = T,
  rescale = T,
  maxiter = 200
)
}
\arguments{
\item{maxiter}{}
}
\description{
Title
}
