% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_last.R
\name{dtms_last}
\alias{dtms_last}
\title{Title}
\usage{
dtms_last(
  matrix,
  transient = NULL,
  time = NULL,
  timestep = NULL,
  dtms = NULL,
  risk,
  risk_to = NULL,
  start_time = NULL,
  start_state = NULL,
  start_distr = NULL,
  end_time = NULL,
  method = "mid",
  sep = "_",
  total = T,
  rescale = T
)
}
\arguments{
\item{rescale}{}
}
\description{
Title
}
