% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_matrix.R
\name{dtms_matrix}
\alias{dtms_matrix}
\title{Title}
\usage{
dtms_matrix(
  transient = NULL,
  absorbing = NULL,
  time = NULL,
  dtms = NULL,
  probs,
  fromvar = "from",
  tovar = "to",
  Pvar = "P",
  enforcedeath = T,
  sep = "_",
  rescale = T
)
}
\arguments{
\item{rescale}{}
}
\description{
Title
}
