% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_probs_summary.R
\name{dtms_probs_summary}
\alias{dtms_probs_summary}
\title{Summarize transition probabilities}
\usage{
dtms_probs_summary(
  probs,
  fromvar = "from",
  tovar = "to",
  timevar = "time",
  probvar = "P",
  sep = "_",
  digits = 4,
  format = "decimal"
)
}
\arguments{
\item{probs}{Data frame with transition probabilities, as generated by 'dtms_transitions'}

\item{fromvar}{Variable with starting state, optional.}

\item{tovar}{Variable with receiving state, optional.}

\item{timevar}{Variable with time scale, optional.}

\item{probvar}{Variable with probabilities, optional.}

\item{sep}{Separator for long state names, default is '_'}

\item{digits}{Number of digits to show, default is 6.}

\item{format}{Show results in decimal format or percentage, default is decimal.}
}
\value{
A data frame
}
\description{
Returns a data frame.
}
\examples{
simple <- dtms(transient=c("A","B"),
               absorbing="X",
               timescale=0:20)
estdata <- dtms_format(data=simpledata,
                       dtms=simple,
                       idvar="id",
                       timevar="time",
                       statevar="state")
estdata <- dtms_clean(data=estdata,
                      dtms=simple)
fit <- dtms_fit(data=estdata)
probs    <- dtms_transitions(dtms=simple,
                             model = fit)
summary(probs)
}
