% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtms_simplify.R
\name{dtms_simplify}
\alias{dtms_simplify}
\title{Simplify state names}
\usage{
dtms_simplify(probs, fromvar = "from", tovar = "to", sep = "_")
}
\arguments{
\item{probs}{Object with transition probabilities}

\item{fromvar}{Name of variable with starting state}

\item{tovar}{Name of variable with receiving state}

\item{sep}{Separator for long state name}
}
\value{
Data frame
}
\description{
Simplify state names
}
